packages/pywhisper/pyproject.toml
[project]
name = "whisper-service"
version = "0.1.0"
description = "FastAPI Whisper-tiny service using openai-whisper"
authors = [
  { name = "Camilla Thorsen", email = "camilla@thorsen.dev" }
]
license = { text = "UNLICENSED" }
requires-python = ">=3.11"

dependencies = [
  "fastapi (>=0.115.12,<0.116.0)",
  "uvicorn[standard] (>=0.34.3,<0.35.0)",
  "python-multipart",
  "openai-whisper",
  "torch (>=2.7.1,<3.0.0)",
  "torchaudio (>=2.7.1,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
